Started work on unistall implementation

+++ ModManager/ModSystem/ModInstaller.cs
@@ -1,9 +1,12 @@
 ﻿using System;
+using System.Collections;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using Modio.Models;
 using ModManager.AddonInstallerSystem;
 using ModManager.AddonSystem;
+using ModManager.MapSystem;
 using ModManager.ModIoSystem;
 using ModManager.PersistenceSystem;
 
@@ -41,14 +44,51 @@ namespace ModManager.ModSystem
 
         public bool Uninstall(Manifest manifest)
         {
+            Console.WriteLine($"inside modinstaller");
+            if (manifest is not Manifest && manifest is MapManifest)
+            {
+                Console.WriteLine($"\"{manifest.ModName} is not a mod.");
+                return false;
+            }
             _installedAddonRepository.Remove(manifest.ModId);
 
+            MarkDllAndManifestFilesForDeletion(manifest);
+
+
+
+            try
+            {
+                //Directory.Delete(manifest.RootPath, true);
+            }
+            catch(UnauthorizedAccessException ex)
+            {
+            }
+            catch(Exception ex)
+            {
+                throw ex;
+            }
             return true;
         }
 
         public bool ChangeVersion(Mod mod, Modio.Models.File file, string zipLocation)
         {
             throw new NotImplementedException();
+        }
+
+        private void MarkDllAndManifestFilesForDeletion(Manifest manifest)
+        {
+            var dirInfo = new DirectoryInfo(manifest.RootPath);
+            var dllFiles = dirInfo.GetFiles("*.dll", SearchOption.AllDirectories);
+            foreach (var dllfile in dllFiles)
+            {
+                dllfile.MoveTo($"dllfile.Name.{Names.Extensions.Remove}");
+            }
+            var manifestFullPath = Path.Combine(manifest.RootPath, Manifest.FileName);
+            var fileInfo = new FileInfo(manifestFullPath);
+            if(fileInfo.Exists)
+            {
+                fileInfo.MoveTo($"{Manifest.FileName}.{Names.Extensions.Remove}");
+            }
         }
     }
 }
